INO: made variable and new command for controlling whether arduino writes to "datalog.txt" or not. This helps to control excessive delays when feedback is not necessary.
HTML: made a button hidden in the advanced view for toggling the writeFile feedback option on and off. Also changed the button view to determine whether white text is needed on some buttons. Also added the code that updates the button text and color to be in the "for loop"