HTML: cosmetic changes to input tags vs a tags for buttons. Also added a function "requestPacer()" to allow the client to request a designated pacer from the arduino Yun.
INO: added code to the feedback command that would be used for this to handle the request and send back the integer that corresponds to the pacer that will be used. This code also secretly sets the lap time to 100 and makes it invisible. To compensate for this, whenever a time is added, the pacer whose time is changing will always be set to visible.