All the changes from Martinsburg and Baltimore trip that I made on my work computer.
changes:
HTML: basically just added another GET request to the requestCommand() function to send data to PHP
updateDBandYun.php: created the file from the updatePacers.php. A lot still needs to be removed, but I added the switch statement for the $command variable which is going to be used to decide which queries to use. Later, I think I will need to add a GET request to the end to send the command to Arduino like what is currently in the index.html requestCommand() function but I will need to remove it from there when it's added to this file.
I think everything with this file is finished now, except for the switch statement to update the database depending on which comomand was received. It does need to be tested though.
It now will update the database for certain commands like backwards, color, and laptime / speed change.
I changed the the database queries format to match how I think it is supposed to be formatted. I couldn't upload it for some reason, so the next step is to upload the changes.
....

The database needs to have the backwards column added and then the backwards section of the switch statement in the updatePacers.php file can be uncommented.

2/1: changed the command in the updateDBandYun.php just by fixing the IP addresses and making sure there were the necessary spaces in the SQL query. and I changed the index.html back to the way it was.

2/2: I'm really starting to lose track of what I've done. I tried to fix the index.html by trying to make the "existing pacers" button work, but I couldn't figure it out.
2/3: I pretty much resolved the issues with using the PIN numbers to access existing pacers and with using the Green button to create a brand new pacer and add times to it and such. I also re-added the "Coach" mode where you can long press on the "Track Practice" header at the top to access a different view where you can control all pacers. However, in coach mode, you can't select each pacer and change it like you could before because it's all dependent on masterPacerSelected being equal to 99. Maybe we can alter that to provide that more complete functionality.
2/4: changed .ino GET Request to include "&newPacerNeeded=1" so it would update the database.

2/5: all /arduino/ GET requests moved from index.html to "updateDBandYun.php"
we should be able to handle all commands in "coach mode" following changes to updateDBandYun.php file
trying to let updatePacers.php correctly update the DB...apparently it was never doing that
updatePacers.php now has an option to retrieve the Pacer data that was already in the database and send it to the Yun when it starts. In order for this to work, you just have to add the "updateYunFromDB()" function to the end of the setup() function in the .ino program.

2/6: everything seems to be working right except for "updateYunFromDB()" in the .INO. However, the PHP request "http://172.20.10.7/sd/TrackPractice/updatePacers.php?pacer=-1&secondsPerLap=-1&needNewPacer=-1&needYunSyncWithDB=1" works correctly. I also observed a problem with entering the PIN "000" being saved as "0" in the database. I also saw a problem with clearing a pacer in the app and the database still showing it as active with a laptime of 1.

2/8: removed "updatePacerDB()" in the .INO. I just moved the call for "updatePacers.php?needYunSyncWithDB=1" to the index.html file and added a button that's available in advanced mode. Upon pressing this button, it brings back all pacers that are still in the database and set to active=1.