No real changes to HTML
updateDBandYun.php: just added ability to handle color changes and speed changes a little bit better. There were previously some logic errors that would cause the lapTime in the DB to be changed when only the color was supposed to be changed. Changing speeds is only available in Coach Mode and it doesn't work perfectly, but it does seem to work well if the coach individually selects the pacer