Added createWorkoutTable.php and runWorkout.php
createWorkoutTable.php: just creates the workout table and inserts a few rows for a very basic workout

runWorkout.php: was able to use this to retrieve workout from the database (workout table) and iterate through the workout while sending the appropriate commands to Arduino. Unfortunately, it looks like PHP will time out after about a minute so it looks like this won't be a viable option, but it does prove the concept and shows some of what is needed.