finished workout feature to add unlimited intervals within each set. It creates an array in the webapp and breaks each array item into separate variables that are sent to PHP and parsed as separate variables, but a very similar process to an array.
It seems to be totally functional except for the problem with delays and running multiple workouts. It could probably be improved by addressing both of those items and the delay that occurs after the user creates the workout. This could probably be done by making all the separate SQL queries into a single SQL query or at least much fewer queries.