Index.html:
I changed the “cancelButton” to look like the “viewEscape” with the text “Back” with an arrow.

I added a developer mode that we can toggle by long-pressing the red button on the initial screen.

I also cleaned up the Extras Menu and added some buttons that are not working (but those are only visible in the developer view)

Added a variable to the runWorkout.php GET request that indicates that the request originated from the app which will allow for more useful feedback messages

On 3/10: added functionality to the button “Change Speed In Flight” in the extras view. It takes the user to a new view where they can change the speed of the pacer while it is in flight.

runWorkout.php:

Added echo statements that make more sense and fixed a glitch that turns the pacer off without any message when the user clicks “Start Workout” without having any workouts already created.

createWorkout.php:

Attempted to change database queries into a single function call by concatenating all the strings of the queries together. I hoped this would reduce the processing time because it takes a long time with lots of intervals, but it didn’t seem to have any effect.