HTML: just changed requestCommand() function to allow for the request to "updateDBandYun" to send commands that are specifically for users as opposed to developers. The Javascript will just determine if the beginning of the string is "USER MSG" and then show the message if it is to the user. But it will show everything when in developer mode.
updateDBandYun.php: added the capability to send messages specifically for users so developer messages could be filtered out. I also fixed the issue with "change speed inflight" not stopping the user from changing to a number outside the range. It now stops them and alerts them if they attempt to do this.