05-11-16 changes in Ohio:
Added labels to the colorSelect page, the lapDistanceSelect page, and the timeInput page.

I added an additional prompt that verifies that the user wants to quit the pacer after they click the quit button.
If they confirm, it returns them to the initial starting page.

When the user selects the distance that they want to put in their time for, the distance that they selected will show in the label on the timeInput page.

Added an alert to display when any help button or link is clicked.